update list api post (not checked for errors)
